# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hand Control API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GestureMapping {
  id: number;
  /** Human-readable gesture name (e.g. "Open Hand", "Pinch") */
  name: string;
  /** Action to trigger (e.g. "rotate_left", "zoom_in", "click") */
  action: string;
  description?: string;
  createdAt: string;
}

export interface CreateGestureBody {
  name: string;
  action: string;
  description?: string;
}

export interface Session {
  id: number;
  label: string;
  gestureCount: number;
  durationSeconds: number;
  createdAt: string;
  endedAt?: string | null;
}

export interface CreateSessionBody {
  label: string;
}

export interface MapperNode {
  id: string;
  size: number;
  color?: string;
  label?: string;
}

export interface MapperEdge {
  source: string;
  target: string;
  weight?: number;
}

export interface MapperGraph {
  nodes: MapperNode[];
  edges: MapperEdge[];
}

export type SessionDetail = Session & {
  mapperGraph?: MapperGraph;
};

export type SessionsSummaryTopGesturesItem = {
  name: string;
  count: number;
};

export interface SessionsSummary {
  totalSessions: number;
  totalGestures: number;
  totalDurationSeconds: number;
  topGestures: SessionsSummaryTopGesturesItem[];
}