# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hand Control API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gestures:
    get:
      operationId: listGestures
      tags: [gestures]
      summary: List gesture mappings
      description: Returns all saved gesture-to-action mappings
      responses:
        "200":
          description: List of gesture mappings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GestureMapping"
    post:
      operationId: createGesture
      tags: [gestures]
      summary: Create a gesture mapping
      description: Save a new gesture-to-action mapping
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGestureBody"
      responses:
        "201":
          description: Created gesture mapping
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GestureMapping"
  /gestures/{id}:
    delete:
      operationId: deleteGesture
      tags: [gestures]
      summary: Delete a gesture mapping
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully
...
components:
  schemas:
...
    GestureMapping:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          description: Human-readable gesture name (e.g. "Open Hand", "Pinch")
        action:
          type: string
          description: Action to trigger (e.g. "rotate_left", "zoom_in", "click")
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - action
        - createdAt
    CreateGestureBody:
      type: object
      properties:
        name:
          type: string
        action:
          type: string
        description:
          type: string
      required:
        - name
        - action
...