# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hand Control API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all saved gesture-to-action mappings
 * @summary List gesture mappings
 */
export const ListGesturesResponseItem = zod.object({
  id: zod.number(),
  name: zod
    .string()
    .describe('Human-readable gesture name (e.g. \"Open Hand\", \"Pinch\")'),
  action: zod
    .string()
    .describe(
      'Action to trigger (e.g. \"rotate_left\", \"zoom_in\", \"click\")',
    ),
  description: zod.string().optional(),
  createdAt: zod.coerce.date(),
});
export const ListGesturesResponse = zod.array(ListGesturesResponseItem);

/**
 * Save a new gesture-to-action mapping
 * @summary Create a gesture mapping
 */
export const CreateGestureBody = zod.object({
  name: zod.string(),
  action: zod.string(),
  description: zod.string().optional(),
});

/**
 * @summary Delete a gesture mapping
 */
export const DeleteGestureParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * Returns all recorded hand tracking sessions
 * @summary List hand tracking sessions
 */
export const ListSessionsResponseItem = zod.object({
  id: zod.number(),
  label: zod.string(),
  gestureCount: zod.number(),
  durationSeconds: zod.number(),
  createdAt: zod.coerce.date(),
  endedAt: zod.coerce.date().nullish(),
});
export const ListSessionsResponse = zod.array(ListSessionsResponseItem);

/**
 * Start a new hand tracking session
 * @summary Create a new session
 */
export const CreateSessionBody = zod.object({
  label: zod.string(),
});

/**
 * @summary Get session details
 */
export const GetSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSessionResponse = zod
  .object({
    id: zod.number(),
    label: zod.string(),
    gestureCount: zod.number(),
    durationSeconds: zod.number(),
    createdAt: zod.coerce.date(),
    endedAt: zod.coerce.date().nullish(),
  })
  .and(
    zod.object({
      mapperGraph: zod
        .object({
          nodes: zod.array(
            zod.object({
              id: zod.string(),
              size: zod.number(),
              color: zod.string().optional(),
              label: zod.string().optional(),
            }),
          ),
          edges: zod.array(
            zod.object({
              source: zod.string(),
              target: zod.string(),
              weight: zod.number().optional(),
            }),
          ),
        })
        .optional(),
    }),
  );

/**
 * Returns aggregate stats across all sessions
 * @summary Get sessions summary stats
 */
export const GetSessionsSummaryResponse = zod.object({
  totalSessions: zod.number(),
  totalGestures: zod.number(),
  totalDurationSeconds: zod.number(),
  topGestures: zod.array(
    zod.object({
      name: zod.string(),
      count: zod.number(),
    }),
  ),
});